Re-split
